First pass ConfigurationHandler
