+++ Slim/App.php
@@ -88,7 +88,11 @@ class App extends \Pimple
     {
         // Settings
         $this['settings'] = $this->share(function ($c) use ($userSettings) {
-            return new \Slim\Configuration($userSettings);
+            $config = new \Slim\Configuration(new \Slim\ConfigurationHandler);
+            $config->setDefaults();
+            $config->setArray($userSettings);
+
+            return $config;
         });
 
         // Environment
