+++ Slim/Configuration.php
@@ -34,7 +34,7 @@ namespace Slim;
 
 /**
  * Configuration
- * Provides app configuration values stored as nested arrays, which can be accessed and stored using dot separated keys.
+ * Uses a ConfigurationHandler class to parse configuration data, accessed as an array.
  *
  * @package    Slim
  * @author     John Porter
@@ -43,22 +43,15 @@ namespace Slim;
 class Configuration implements \ArrayAccess, \IteratorAggregate
 {
     /**
-     * Cache of previously parsed keys
-     * @var array
+     * Handler for Configuration values
+     * @var mixed
      */
-    protected $keys = array();
-
+    protected $handler;
     /**
      * Storage array of values
      * @var array
      */
     protected $values = array();
-
-    /**
-     * Expected nested key separator
-     * @var string
-     */
-    protected $separator = '.';
 
     /**
      * Default values
@@ -90,12 +83,33 @@ class Configuration implements \ArrayAccess, \IteratorAggregate
 
     /**
      * Constructor
-     * Merge provided values with the defaults to ensure all required values are set
-     * @param array $values
+     * @param mixed $handler
      */
-    public function __construct(array $values = array())
+    public function __construct(\Slim\ConfigurationHandler $handler)
     {
-        $this->values = $this->mergeArrays($this->getDefaults(), $values);
+        if (!$handler instanceof \ArrayAccess) {
+            throw new \InvalidArgumentException("ConfigurationHandler must implement \\ArrayAccess.");
+        }
+
+        $this->handler = $handler;
+
+        $this->setDefaults();
+    }
+
+    /**
+     * Set Slim's defaults using the handler
+     */
+    public function setArray(array $values)
+    {
+        $this->handler->setArray($values);
+    }
+
+    /**
+     * Set Slim's defaults using the handler
+     */
+    public function setDefaults()
+    {
+        $this->handler->setArray($this->defaults);
     }
 
     /**
@@ -108,184 +122,34 @@ class Configuration implements \ArrayAccess, \IteratorAggregate
     }
 
     /**
-     * Get all values as nested array
-     * @return array
-     */
-    public function getAllNested()
-    {
-        return $this->values;
-    }
-
-    /**
-     * Get all values as flattened key array
-     * @return array
-     */
-    public function getAllFlat()
-    {
-        return $this->flattenArray($this->values);
-    }
-
-    /**
-     * Get all flattened array keys
-     * @return array
-     */
-    public function getKeys()
-    {
-        $flattened = $this->flattenArray($this->values);
-        return array_keys($flattened);
-    }
-
-    /**
-     * Get a value from a nested array based on a separated key
-     * @param  string $key
+     * Call a method from the handler
+     * @param  string $method
+     * @param  array $params
      * @return mixed
      */
+    public function callHandlerMethod($method, array $params = array())
+    {
+        return call_user_func_array(array($this->handler, $method), $params);
+    }
+
     public function offsetGet($key)
     {
-        return $this->getValue($key, $this->values);
+        return $this->handler[$key];
     }
 
-    /**
-     * Set nested array values based on a separated key
-     * @param  string  $key
-     * @param  mixed   $value
-     * @return array
-     */
     public function offsetSet($key, $value)
     {
-        $this->setValue($key, $value, $this->values);
+        $this->handler[$key] = $value;
     }
 
-    /**
-     * Check an array has a value based on a separated key
-     * @param  string  $key
-     * @return boolean
-     */
     public function offsetExists($key)
     {
-        return (bool)$this->getValue($key, $this->values);
+        return isset($this->handler[$key]);
     }
 
-    /**
-     * Remove nested array value based on a separated key
-     * @param  string  $key
-     */
     public function offsetUnset($key)
     {
-        $keys = $this->parseKey($key);
-        $array = &$this->values;
-
-        while (count($keys) > 1) {
-            $key = array_shift($keys);
-
-            if (!isset($array[$key]) || !is_array($array[$key])) {
-                return;
-            }
-
-            $array =& $array[$key];
-        }
-
-        unset($array[array_shift($keys)]);
-    }
-
-    /**
-     * Parse a separated key and cache the result
-     * @param  string $key
-     * @return array
-     */
-    protected function parseKey($key)
-    {
-        if (!isset($this->keys[$key])) {
-            $this->keys[$key] = explode($this->separator, $key);
-        }
-
-        return $this->keys[$key];
-    }
-
-    /**
-     * Get a value from a nested array based on a separated key
-     * @param  string $key
-     * @param  array  $array
-     * @return mixed
-     */
-    protected function getValue($key, array $array = array())
-    {
-        $keys = $this->parseKey($key);
-
-        while (count($keys) > 0 and !is_null($array)) {
-            $key = array_shift($keys);
-            $array = isset($array[$key]) ? $array[$key] : null;
-        }
-
-        return $array;
-    }
-
-    /**
-     * Set nested array values based on a separated key
-     *
-     * @param  string  $key
-     * @param  mixed   $value
-     * @param  array   $array
-     * @return array
-     */
-    protected function setValue($key, $value, array &$array = array())
-    {
-        $keys = $this->parseKey($key, $this->separator);
-        $pointer = &$array;
-
-        while (count($keys) > 0) {
-            $key = array_shift($keys);
-            $pointer[$key] = (isset($pointer[$key]) ? $pointer[$key] : array());
-            $pointer = &$pointer[$key];
-        }
-
-        $pointer = $value;
-        return $array;
-    }
-
-    /**
-     * Merge arrays with nested keys
-     * Usage: $this->mergeArrays(array $array [, array $...])
-     * @return array
-     */
-    protected function mergeArrays()
-    {
-        $arrays = func_get_args();
-        $merged = array();
-
-        foreach ($arrays as $array) {
-            foreach ($array as $key => $value) {
-                $merged = $this->setValue($key, $value, $merged);
-            }
-        }
-
-        return $merged;
-    }
-
-    /**
-     * Flatten a nested array to a separated key
-     *
-     * @param  array  $array
-     * @param  string $separator
-     * @return array
-     */
-    protected function flattenArray(array $array, $separator = null)
-    {
-        $flattened = array();
-
-        if (is_null($separator)) {
-            $separator = $this->separator;
-        }
-
-        foreach ($array as $key => $value) {
-            if (is_array($value)) {
-                $flattened = array_merge($flattened, $this->flattenArray($value, $key.$separator));
-            } else {
-                $flattened[trim($separator.$key, $this->separator)] = $value;
-            }
-        }
-
-        return $flattened;
+        unset($this->handler[$key]);
     }
 
     /**
